fix(config): enable balance overrides on Lisk, don't check relayer balance in healthchecks
We need to enable balance overrides because there are clients who are calling
eth_estimateUserOperationGas with empty accounts before actually calling the paymaster for
sponsorship. It was previously disabled because it didn't work on the lisk rpc, but now it works.
Also don't check relayer balance in healthchecks, just check the master account and assume if it's
funded it will fund the relayers. Otherwise we encountered some weird errors that would cause the
checks to fail because of problems with relayer balance accounting

BREAKING CHANGE: We are not checking the relayer balance in healtchecks any more, by default. It can
be enabled in production config (if necessary).